==> 1 Web Scraper.ipynb <==
# Web Scraper file

The data would be extracted from The Real Estate website: PropertyPro.NG

I have researched the website and see some features that most houses/ listings have:

After that for each city we would get the following information
* Title of the home/listing - which we can get info like
    * If the house is a duplex, apartment, land, flat
___
* Location of home : we can get info like
    * city home is located, neighborhood/area it is located
___
* Price of Home : It is a string and would be converted into integers
___
* If home is Newly Built (True/ False)
* If home is Serviced (True/ False)
* If home is Furnished (True/ False)
___
* No of Bedrooms (String convert to integer)
___
* No of Bathrooms (String convert to integer)
___
* No of Toilets (String convert to integer)
___
* No of Bathrooms (String convert to integer)

# !pip install bs4 requests --quiet

# Import Seperate Libraries
from bs4 import BeautifulSoup
import requests
import csv

from requests.adapters import HTTPAdapter
from requests.packages.urllib3.util.retry import Retry

# This would make us scrape in sessions so that if we have network issues
# we wont need to start scraping again we would just revert to the previous
# session tracked
session = requests.Session()
retry = Retry(connect=3, backoff_factor=0.5)
adapter = HTTPAdapter(max_retries=retry)
session.mount('http://', adapter)
session.mount('https://', adapter)

# Save in this file
csv_file = open("name of city.csv", "w")

csv_writer = csv.writer(csv_file)

csv_writer.writerow(["Title", "Location", "Price","Serviced", "Newly Built", "Furnished", "Bedrooms", "Bathrooms", "Toilets"])

# Get Location data
for page in range(0, 484):
    # Replace the ikoyi with the city you want e.g. surulere, ajah, ikeja, lekki
    link = session.get(f"https://www.propertypro.ng/property-for-rent/in/lagos?page={page}").text 
    soup = BeautifulSoup(link, "lxml")
    
    for listing in soup.find_all("div", class_="single-room-sale listings-property"):
        title = listing.find("h3", class_="listings-property-title2").text
        location = listing.find_all("h4")[1].text

        price_info = listing.find("h3", class_="listings-price")
        price = price_info.find_all("span")[1].text

        if listing.find("a", href="/property-for-sale/is-new"):
          is_new = 1
        else:
          is_new = 0
        
        if listing.find("a", href="/property-for-sale/is-serviced"):
          serviced = 1
        else:
          serviced = 0
        
        if listing.find("a", href="/property-for-sale/is-furnished"):
          furnished = 1
        else:
          furnished = 0
          
        more_info = listing.find("div", class_="fur-areea")
        bed = more_info.find_all("span")[0].text
        bathroom = more_info.find_all("span")[1].text
        toilet = more_info.find_all("span")[2].text

        csv_writer.writerow([title, location, price, serviced, is_new, furnished, bed, bathroom, toilet])
print("Done")

csv_file.close()
