==> 2 Lagos Extracting and Cleaning.ipynb <==
import pandas as pd

rent = pd.read_csv("data/nigeria-rent.csv")

rent

## Choosing Rented Residentials

rent["More Info"].sample(5)

def rental(feature):
    final = []
    for row in feature:
        row = row.lower()
        if ("bedroom" in row) or ("flat" in row) or ("house" in row) or ("apartment" in row) or ("self contain" in row) and (not "office" in row):
            final.append("rent")
        else:
            final.append("other.row")
    return final

rent["Type"] = rental(rent["More Info"])

rent = rent[rent["Type"] == "rent"].copy()

## Extracting State

rent["Location"]

def state(feature):
    return [row.split(" ")[-1] for row in feature]

rent["State"] = state(rent["Location"])

lagos = rent[rent["State"] == "Lagos"]

## Extract Lagos Cities

lagos.head()

# Drop duplicates
lagos.drop_duplicates(inplace=True)

def city(feature):
    return [row.split(" ")[-2] for row in feature]

lagos["City"] = city(lagos["Location"])

lagos["City"].value_counts(normalize=True).head(10)

lagos = lagos[(lagos["City"] == "Lekki") | (lagos["City"] == "Ajah") | (lagos["City"] == "Yaba") | (lagos["City"] == "Ojodu") | (lagos["City"] == "Ikoyi") |
(lagos["City"] == "Ikeja") | (lagos["City"] == "Surulere") | (lagos["City"] == "Gbagada") | (lagos["City"] == "Island")].copy()

lagos["City"].value_counts()

## Extracting Lagos Neighborhood

### Lekki

lekki = lagos[lagos["City"] == "Lekki"]

final = []
for row in lekki["Location"]:
    try:
        final.append(row.split(" ")[-3])
    except:
        final.append("Other")
lekki["Neighborhood"] = final        

final = []
for value in lekki["Location"]:
    if "1" in value:
        final.append("Lekki Phase 1")
    elif "chevron" in value.lower():
        final.append("Chevron")
    elif "ikate" in value.lower():
        final.append("Ikate")
    elif "ikota" in value.lower():
        final.append("Ikota")
    elif ("london" in value.lower()) or ("osapa" in value.lower()):
        final.append("Osapa London")
    elif "agungi" in value.lower():
        final.append("Agungi")
    elif "ologolo" in value.lower():
        final.append("Ologolo")
    else:
        final.append("Other Lekki")

lekki["Neighborhood"] = final

lekki["Neighborhood"].value_counts()

### Ajah

ajah = lagos[lagos["City"] == "Ajah"]

final = []
for row in ajah["Location"]:
    try:
        final.append(row.split(" ")[-3])
    except:
        final.append("Other")
ajah["Neighborhood"] = final        

final = []
for value in ajah["Location"]:
    if "badore" in value.lower():
        final.append("Badore")
    elif "sangotedo" in value.lower():
        final.append("Sangotedo")
    elif ("ado" in value.lower()) and not ("badore" in value.lower()):
        final.append("Ado")
    elif "lekki" in value.lower():
        final.append("Lekki Scheme")
    else:
        final.append("Other Ajah")

ajah["Neighborhood"] = final
ajah["Neighborhood"].value_counts()

### Yaba

yaba = lagos[lagos["City"] == "Yaba"]

final = []
for row in yaba["Location"]:
    try:
        final.append(row.split(" ")[-3])
    except:
        final.append("Other")
yaba["Neighborhood"] = final        

final = []
for value in yaba["Location"]:
    if "akoka" in value.lower():
        final.append("Akoka")
    elif "alagomeji" in value.lower():
        final.append("Alagomeji")
    elif "onike" in value.lower():
        final.append("Onike")
    elif "agoro" in value.lower():
        final.append("Agoro")
    elif "metta" in value.lower():
        final.append("Ebute-Metta")
    elif "sabo" in value.lower():
        final.append("Sabo")
    elif "jibowu" in value.lower():
        final.append("Jibowu")
    elif "ijesha" in value.lower():
        final.append("Ijesha")
    else:
        final.append("Other Yaba")
yaba["Neighborhood"] = final
yaba["Neighborhood"].value_counts()

### Ojodu

ojodu = lagos[lagos["City"] == "Ojodu"]

final = []
for row in ojodu["Location"]:
    try:
        final.append(row.split(" ")[-3])
    except:
        final.append("Other")
ojodu["Neighborhood"] = final        

final = []
for value in ojodu["Location"]:
    if ("magodo" in value.lower()) and ("1" in value.lower()):
        final.append("Magodo Gra Phase 1")
    elif ("omole" in value.lower()) and ("2" in value.lower()):
        final.append("Omole Phase 2")
    elif "north" in value.lower():
        final.append("Isheri North")
    elif "olowora" in value.lower():
        final.append("Olowora")
    elif ("omole" in value.lower()) and ("1" in value.lower()):
        final.append("Omole Phase 1")
    elif ("berger" in value.lower()):
        final.append("Ojodu Berger")
    else:
        final.append("Other Ojodu")
ojodu["Neighborhood"] = final
ojodu["Neighborhood"].value_counts()

### Ikoyi

ikoyi = lagos[lagos["City"] == "Ikoyi"]

final = []
for row in ikoyi["Location"]:
    try:
        final.append(row.split(" ")[-3])
    except:
        final.append("Other")
ikoyi["Neighborhood"] = final        

final = []
for value in ikoyi["Location"]:
    if "banana island" in value.lower():
        final.append("Banana Island")
    elif "old ikoyi" in value.lower():
        final.append("Old Ikoyi")
    elif "parkview estate" in value.lower():
        final.append("Parkview Estate")
    elif "dolphin estate" in value.lower():
        final.append("Dolphin Estate")
    elif "osborn" in value.lower():
        final.append("Osborne Foreshore Estate")
    elif "bourdillon" in value.lower():
        final.append("Bourdillon")
    elif "awolowo road" in value.lower():
        final.append("Awolowo Road")
    else:
        final.append("Other Ikoyi")
ikoyi["Neighborhood"] = final
ikoyi["Neighborhood"].value_counts()

### Ikeja

ikeja = lagos[lagos["City"] == "Ikeja"]

final = []
for row in ikeja["Location"]:
    try:
        final.append(row.split(" ")[-3])
    except:
        final.append("Other")
ikeja["Neighborhood"] = final        

final = []
for value in ikeja["Location"]:
    if "gra" in value.lower():
        final.append("GRA")
    elif "jones" in value.lower():
        final.append("Adeniyi Jones")
    elif "avenue" in value.lower():
        final.append("Allen Avenue")
    elif "opebi" in value.lower():
        final.append("Opebi")
    elif "oregun" in value.lower():
        final.append("Oregun")
    elif "alausa" in value.lower():
        final.append("alausa")
    elif "awolowo" in value.lower():
        final.append("Awolowo way")
    elif "toyin" in value.lower():
        final.append("Toyin street")
    else:
        final.append("Other Ikeja")
ikeja["Neighborhood"] = final
ikeja["Neighborhood"].value_counts()

### Surulere

surulere = lagos[lagos["City"] == "Surulere"]

final = []
for row in surulere["Location"]:
    try:
        final.append(row.split(" ")[-3])
    except:
        final.append("Other")
surulere["Neighborhood"] = final        

final = []
for value in surulere["Location"]:
    if "aguda" in value.lower():
        final.append("Aguda")
    elif "thomas" in value.lower():
        final.append("Bode thomas")
    elif "marsha" in value.lower():
        final.append("Marsha")
    elif "ogunlana" in value.lower():
        final.append("Ogunlana")
    elif "lawanson" in value.lower():
        final.append("Lawanson")
    elif "adelabu" in value.lower():
        final.append("Adelabu")
    elif "avenue" in value.lower():
        final.append("Randle Avenue")
    elif "ijesha" in value.lower():
        final.append("Ijesha")
    elif "ojuelegba" in value.lower():
        final.append("Ojuelegba")
    # elif "araba" in value.lower():
    #     final.append("Idi Araba")
    else:
        final.append("Other Surulere")


surulere["Neighborhood"] = final
surulere["Neighborhood"].value_counts()

### Island

island = lagos[lagos["City"] == "Island"]

final = []
for row in island["Location"]:
    try:
        final.append(row.split(" ")[-4])
    except:
        final.append("Other")
island["Neighborhood"] = final        

final = []
for value in island["Location"]:
    if "oniru" in value.lower():
        final.append("Oniru")
    elif "extension" in value.lower():
        final.append("Victoria Island Extension")
    elif "odeku" in value.lower():
        final.append("Adeola Odeku")
    elif "1004" in value.lower():
        final.append("1004")
    elif "ayorinde" in value.lower():
        final.append("Ligali Ayorinde")
    else:
        final.append("Other Victoria Island")
island["Neighborhood"] = final
island["Neighborhood"].value_counts()

### Gbagada

gbagada = lagos[lagos["City"] == "Gbagada"]

final = []
for row in gbagada["Location"]:
    try:
        final.append(row.split(" ")[-3])
    except:
        final.append("Other")
gbagada["Neighborhood"] = final        

final = []
for value in gbagada["Location"]:
    if "ifako" in value.lower():
        final.append("Ifako")
    elif "1" in value.lower():
        final.append("Gbagada Phase 1")
    elif "2" in value.lower():
        final.append("Gbagada Phase 2")
    elif "soluyi" in value.lower():
        final.append("Soluyi")
    elif "oworonshoki" in value.lower():
        final.append("Oworonshoki")
    elif "ups" in value.lower():
        final.append("Millennium Ups")
    elif "medina" in value.lower():
        final.append("Medina")
    else:
        final.append("Other Gbagada")


gbagada["Neighborhood"] = final
gbagada["Neighborhood"].value_counts()

lagos = pd.concat([lekki, ikoyi, ikeja, island, ojodu, gbagada, surulere, ajah, yaba]).drop(["Location","State", "Type"], axis=1)

lagos.to_csv("data/lagos-rent.csv", index=False)
