==> 3 Data Preprocessing.ipynb <==
import pandas as pd
import joblib

lagos = pd.read_csv("data/lagos-rent.csv")

lagos.head()

lagos.info()

def house_type(serie):
    final = []
    
    # Extract property name from title
    for title in list(serie):
        # convert all titles to lowercase trings
        title = str(title).lower()
        if ("land" in title) and (not "island" in title):
            final.append("land")
        elif ("flat" in title) and (not "mini" in title):
            final.append("Flat")
        elif ("self" in title) or ("mini" in title):
            final.append("Self-contained")
        elif ("apartment" in title) or ("tment" in title):
            final.append("Apartment")
        elif "semi" in title:
            final.append("Semi Detached Duplex")
        elif "terra" in title:
            final.append("Terraced Duplex")
        elif ('fully detached' in title) or ('bedroom detached duplex' in title) and (not "semi" in title):
            final.append("Fully Detached Duplex")
        elif ('bedroom' in title) or ('bed room' in title) and (not "semi" in title) and (not "terra" in title) and (not "apartment" in title) and (not "flat" in title):
            final.append("Fully Detached Duplex")
        elif ('detached' in title)  and (not "semi" in title)  and (not "terra" in title):
            final.append("Fully Detached Duplex")
        elif ('duplex' in title)  and (not "semi" in title)  and (not "terra" in title):
            final.append("Fully Detached Duplex")
        else:
            final.append("others")
    return final

lagos["Property Type"] = house_type(lagos["Title"])

lagos["Property Type"].value_counts()

mask = lagos.loc[(lagos["Property Type"]=="others") | (lagos["Property Type"]=="land")].index
lagos.drop(mask, inplace=True)
lagos.drop(["Title", "More Info"], axis=1, inplace=True)

lagos["Property Type"].value_counts()

# Remove the last item from the price
def price_split(feature):
    final = []
    for value in feature:
        # if value has a price period, extract it else let it's price period be fixed
        try:
            final.append(value.split("/")[1])
        except:
            final.append("year")
    return final

# applying function
lagos["Price Period"] = price_split(lagos["Price"])

def comma_remove(feature):
    final = []
    for sample in feature:
        # Get the prices
        sample = sample.split("/")[0]
        
        # Remove the commas
        if "," in sample:
            final.append(sample.replace(",",""))
        else:
            final.append(sample)
    
    # Convert them to numbers
    int_final = [float(item) for item in final]
    return int_final


# applying functions
lagos["Price"] = comma_remove(lagos["Price"])

lagos = lagos[lagos["Price Period"] == "year"].copy()
lagos.drop("Price Period", axis=1, inplace=True)

def remove_bed_strip(serie):
    final = []
    # If value doesn't have a number, replace it with a zero,
    # else replace it with the number it contains
    for value in serie:
        if value.split("beds")[0].strip() == "":
            final.append(int("0"))
        else:
            final.append(int(value.split("beds")[0].strip()))
    return final

def remove_toilet_strip(serie):
    final = []
    for value in serie:
        if value.split("Toilets")[0].strip() == "":
            final.append(int("0"))
        else:
            final.append(int(value.split("Toilets")[0].strip()))
    return final 

def remove_bath_strip(serie):
    final = []
    for value in serie:
        if value.split("baths")[0].strip() == "":
            final.append(int("0"))
        else:
            final.append(int(value.split("baths")[0].strip()))
    return final 

Applying functions to three features

lagos["Bedrooms"] = remove_bed_strip(lagos["Bedrooms"])

lagos["Bathrooms"] = remove_bath_strip(lagos["Bathrooms"])

lagos["Toilets"] = remove_toilet_strip(lagos["Toilets"])

lagos[["Bedrooms","Bathrooms","Toilets"]].sample(5)

def first_impute(f1, f2):
    final = []
    for value in zip(f1, f2):
        if (value[0] == 0):
            final.append(value[1])
        else:
            final.append(value[0])
    return final

def second_impute(f1, f2):
    final = []
    for value in zip(f1, f2):
        if value[0] > 1 :
            if (value[0] == 0):
                final.append(value[1])
            else:
                final.append(value[0])
        else:
            if (value[0] == 0):
                final.append(1)
            else:
                final.append(value[0])
    return final

lagos.shape

lagos["Bedrooms"] = first_impute(lagos["Bedrooms"], lagos["Bathrooms"])
lagos["Bathrooms"] = first_impute(lagos["Bathrooms"], lagos["Bedrooms"])
lagos["Toilets"] = second_impute(lagos["Toilets"], lagos["Bathrooms"])

# try missing room imputation again here
mask = lagos.loc[(lagos['Bedrooms'] == 0) | (lagos['Bedrooms'] == 0) | (lagos['Toilets'] == 0)].index
lagos.drop(mask, inplace=True)


from feature_engine.outliers import Winsorizer

lagos["Price"].median() /1e6, lagos["Price"].mean()/1e6

# # try price outlier handling and missing room imputation again here
mask = lagos.loc[(lagos["Price"] < 1e5) | (lagos["Price"] > 5e7)].index
lagos.drop(mask, inplace=True)

lagos["Price"].median() /1e6, lagos["Price"].mean()/1e6

## Data Preprocessing

#### Data Split

Splitting our data before data preprocessing to avoid data leakage

# Spliting to extract 20% test data
from sklearn.model_selection import StratifiedShuffleSplit
split = StratifiedShuffleSplit(n_splits=1, test_size=0.2, random_state=42)
for train_index, test_index in split.split(lagos, lagos["Neighborhood"]):
    train_set = lagos.iloc[train_index]
    test_set = lagos.iloc[test_index]

lagos = train_set.drop("Price", axis=1)
price = train_set["Price"]

Let's seperate our numerical columns and our categorical ones

categorical_col = [column for column in lagos.columns if lagos[column].dtype == "O"]

lagos[categorical_col].sample(4)

lagos = lagos.sample(frac=1, random_state=51)

#### Feature Encoding

We need to perform OneHotEncoding on some of the norminal categorical variables (variables with no order e.g. Nigeria, Congo, Ghana)

# import library for encoding
from feature_engine.encoding import OneHotEncoder

encoder = OneHotEncoder()
encoder.fit(lagos)
lagos = encoder.transform(lagos)

Saving Encoder for future use in model deployment

joblib.dump(encoder, "tools/encoder_joblib")

lagos

Save preprocessed data for Model Building

lagos = pd.concat([lagos, price], axis=1)

lagos.to_csv('data/train.csv', index=False)
test_set.to_csv('data/test.csv', index=False)
