==> 4 Model Building.ipynb <==
import pandas as pd
import numpy as np
import joblib

Importing train, test and validation dataset

train = pd.read_csv("data/train.csv")
test = pd.read_csv("data/test.csv")

train.head()

train.shape, test.shape

Let's do the necessary splits

X_train = train.drop(["Price"], axis=1)
y_train = train["Price"]

Here are the categorical and numerical features that need to be preprocessed

Importing tools needed for preprocessing

encoder = joblib.load("tools/encoder_joblib")

Splitting test data

X_test = test.drop(["Price"], axis=1)
y_test = test["Price"]

Here we are preprocessing the test data so it can be used by the machine learning model

X_test = encoder.transform(X_test) # enconding categorical values

X_test.head()

## Model Building and Validation

Lets import the necessary machinne learning models that would be used for fitting the data

from sklearn.ensemble import RandomForestRegressor # Random Forest
from xgboost import XGBRegressor # XGBoost
from sklearn.svm import SVR # SVM
from sklearn.neighbors import KNeighborsRegressor # KNN
from lightgbm import LGBMRegressor # LightGBM

### Random Forest

Here we are initializing the model, then fitting the data and making prediction on the validation data

rf = RandomForestRegressor(max_depth=11, n_estimators=20, random_state=51)
rf.fit(X_train, y_train)
rf_predictions = rf.predict(X_test)

### XGBoost

xgb = XGBRegressor(random_state=51, n_estimators=60, learning_rate=0.202)
xgb.fit(X_train, y_train)
xgb_predictions = xgb.predict(X_test)

### LightGBM

# lgb = LGBMRegressor(random_state=51)
lgb = LGBMRegressor(random_state=51, n_estimators=117)
lgb.fit(X_train, y_train)
lgb_predictions = lgb.predict(X_test)

### Model Performance on Test data

Here we are checking the performance of the machine learning models on the test data

# import MAE metric
from sklearn.metrics import mean_absolute_error

print(f"Random Forest: {mean_absolute_error(y_test, rf_predictions) / 1e6}")
print(f"XGBoost: {mean_absolute_error(y_test, xgb_predictions) / 1e6}")
print(f"LightGBM: {mean_absolute_error(y_test, lgb_predictions) / 1e6}")

### Model Save for Deployment

joblib.dump(lgb, "tools/model_joblib")
